refactor(overlay): unify linux x11/wayland overlays behind surface primitives
The X11 and wlroots overlay backends were ~82% copy-paste: every
drawing, layout, offset and animation method existed twice (2,263
lines total) and the copies had already begun to drift.

All shared logic now lives once, driving a small primitive interface
that each backend implements: HiDPI scale, buffer lifecycle
(prepare / acquire-or-drop-frame / clear variants), flush/hide, the
click-indicator map/unmap pair, a pre-animation sync hook, and four
cgo draw calls. The backends shrink to their genuine differences —
X11 (1,072 -> 375 lines): the Xft.dpi scale probe, the double-buffered
animation clear, a single always-available drawing target; wlroots
(1,193 -> 497): shared-memory buffer acquisition with one sync-retry
and frame-drop, the keyboard poller lifecycle, and layer-surface
keyboard-capture handling. Net -1,393 lines.

Behavioral quirks are preserved deliberately: the animation
goroutine's deadlock guard (re-check stop after acquiring the render
lock), Wayland's no-clear-on-an-unselected-buffer indicator teardown
vs X11's clear+hide, badge paths that prepare buffers without
selecting a frame, and the X11-only HiDPI scale — now applied
centrally in the shared draw helpers instead of at every call site.

Exported backend methods stay thin nil-guarded delegates rather than
promoted methods: a promoted method called through a nil backend
pointer panics before any receiver guard runs, and the manager relies
on nil-safe calls. The manager, both no-cgo stubs, and the stub
contract test are unchanged.
